CSQC: pass 62-bit sendflags to PR2 SendEntity as lo/hi ints (F2)

+++ src/pr2.h
@@ -91,7 +91,7 @@ void        PR2_ClearEdict(edict_t* e);
 #define PR_ClearEdict PR2_ClearEdict
 
 #ifdef FTE_PEXT_CSQC
-qbool       PR2_SendEntity(edict_t* e, edict_t* to, int sendflags);
+qbool       PR2_SendEntity(edict_t* e, edict_t* to, uint64_t sendflags);
 void        PR2_QCRequest(edict_t* cl_ent, int argcount);
 #endif
